fix: use bracket notation for hyphenated jsonpath keys and extract shared metadata helpers
Agent-Logs-Url: https://github.com/lakehq/sail/sessions/f12cf27f-4948-4b1f-a78b-982995cffc3f
