+++ crates/sail-iceberg/src/utils/mod.rs
@@ -11,6 +11,7 @@
 // limitations under the License.
 
 pub mod conversions;
+pub mod metadata;
 pub mod partition_transform;
 pub mod snapshot_id;
 pub mod timestamp;
